LayoutTests/fast/mediastream/MediaStreamTrackEvent-constructor.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
    </head>
        <script>

            var mediaStream = null;
            var mediaStreamTrack = null;
            var timer = null;

            window.jsTestIsAsync = true;

            description("This tests the constructor for the MediaStreamTrackEvent DOM class.");

            function testMediaStreamTrackEvent()
            {
                debug("<br>*** No initializer passed ***");
                shouldBe("new MediaStreamTrackEvent('MediaStreamTrackEvent').bubbles", "false");
                shouldBe("new MediaStreamTrackEvent('MediaStreamTrackEvent').cancelable", "false");
                shouldBeNull("new MediaStreamTrackEvent('MediaStreamTrackEvent').track");
    
                debug("<br>*** Bubbles and cancelable true, track is missing ***");
                shouldBe("new MediaStreamTrackEvent('MediaStreamTrackEvent', { bubbles: true, cancelable: true }).bubbles", "true");
                shouldBe("new MediaStreamTrackEvent('MediaStreamTrackEvent', { bubbles: true, cancelable: true }).cancelable", "true");
                shouldBeNull("new MediaStreamTrackEvent('MediaStreamTrackEvent', { bubbles: true, cancelable: true }).track");
    
                debug("<br>*** Bubbles and cancelable true, invalid track ***");
                shouldBe("new MediaStreamTrackEvent('MediaStreamTrackEvent', { bubbles: true, cancelable: true, track: MediaStreamTrack }).bubbles", "true");
                shouldBe("new MediaStreamTrackEvent('MediaStreamTrackEvent', { bubbles: true, cancelable: true, track: MediaStreamTrack }).cancelable", "true");
                shouldBeNull("new MediaStreamTrackEvent('MediaStreamTrackEvent', { bubbles: true, cancelable: true, track: MediaStreamTrack }).track");
    
                debug("<br>*** Initialize 'track' with a invalid values ***");
                shouldBeNull("new MediaStreamTrackEvent('MediaStreamTrackEvent', { track: 10 }).track");
                shouldBeNull("new MediaStreamTrackEvent('MediaStreamTrackEvent', { track: \'string\' }).track");
                emptyObject = { };
                shouldBeNull("new MediaStreamTrackEvent('MediaStreamTrackEvent', { track: emptyObject }).track");
                shouldBeNull("new MediaStreamTrackEvent('MediaStreamTrackEvent', { track: document }).track");

                debug("<br>*** Bubbles and cancelable true, valid track ***");
                shouldBe("new MediaStreamTrackEvent('MediaStreamTrackEvent', { bubbles: true, cancelable: true, track: mediaStreamTrack }).bubbles", "true");
                shouldBe("new MediaStreamTrackEvent('MediaStreamTrackEvent', { bubbles: true, cancelable: true, track: mediaStreamTrack }).cancelable", "true");

                debug("<br>*** Initialize 'track' with valid track object ***");
                shouldBeNonNull("new MediaStreamTrackEvent('MediaStreamTrackEvent', { track: mediaStreamTrack }).track");
                shouldBe("new MediaStreamTrackEvent('MediaStreamTrackEvent', { track: mediaStreamTrack }).track", "mediaStreamTrack");

                debug("");
                finishJSTest();
            }

            function gotStream(stream)
            {
                clearTimeout(timer);
                testPassed('getUserMedia succeeded.');

                mediaStream = stream;
                shouldBeNonNull("mediaStream");
                shouldBeNonNull("mediaStream.getVideoTracks()[0]");
                mediaStreamTrack = mediaStream.getVideoTracks()[0];
                testMediaStreamTrackEvent();
            }

            function primeTimeout(msg)
            {
                clearTimeout(timer);
                timer = setTimeout(function() { testFailed(msg); finishJSTest() }, 4000);
            }

            function test()
            {
            /* 
                @@@@@ NOTE: we can't get a MediaStreamTrack until we have a functional mock getUserMedia, so just test what we can for now @@@@@ 
                primeTimeout("Timeout waiting for getUserMedia()");
                shouldNotThrow('getUserMedia({audio:true, video:true}, gotStream)');
            */

                testMediaStreamTrackEvent();
            }

        </script>
        <script src="../../resources/js-test-post.js"></script>

    <body onload="test()">
    </body>
</html>
